+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # UNRELEASED
 
+**BREAKING CHANGES**
+
+-   Rename `DecryptionErrorCode.SenderIdentityPreviouslyVerified` to
+    `SenderIdentityVerificationViolation` (in line with changes to
+    matrix-rust-sdk).
+
+-   Rename `UserIdentity` to `OtherUserIdentity` (in line with changes
+    to matrix-rust-sdk).
+
+-   Update matrix-rust-sdk to `#5d46b35d`.
+
 # matrix-sdk-crypto-wasm v9.1.0
 
 -   Update matrix-rust-sdk to `866b6e5f`, which includes: